new query manager mixins for routes + components
These mixins are intended to replace most direct usage of the Apollo
service. The query managers act as a proxy to the Apollo service,
tracking all watched queries made in a route or component and
automatically unsubscribing from them via the respective lifecycle
hooks.

Additionally, the objects returned by watchQuery include an
`_apolloObservable` property that gives direct access to the
ObservableQuery returned by ApolloClient.watchQuery. This should make it
easy to refetch queries or handle paginated result sets.

This is poorly tested so far!